Fixing crashing bug with favorites
